requirements.txt
# requirements.txt

# MCP SDK
mcp>=0.9.0

# Web frameworks
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
aiohttp>=3.9.0

# Security tools integration
dnspython>=2.4.2
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.3

# Database
sqlalchemy>=2.0.0
asyncpg>=0.29.0
redis>=5.0.0

# Utilities
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
loguru>=0.7.2
python-json-logger>=2.0.7

# Security & Crypto
cryptography>=41.0.7
pyotp>=2.9.0
pyjwt>=2.8.0

# Reporting
jinja2>=3.1.2
markdown>=3.5.0
reportlab>=4.0.7

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
httpx>=0.25.2